파일 다운로드 작업까지 커밋

+++ Board/src/main/java/com/study/domain/post/AttachMapper.java
@@ -9,7 +9,7 @@ public interface AttachMapper {
 
 	public int insertAttach(List<AttachDto> attachList);
 	// 파일 다운로드에서 사용
-	public AttachDto selectAttachDetail(Long id);
+	public AttachDto selectAttachDetail(int id);
 
 	public int deleteAttach(int boardIdx);
